LayoutTests/imported/w3c/web-platform-tests/content-security-policy/script-src/worker-data-set-timeout.sub-expected.txt


FAIL Shared worker with data: url inherits CSP assert_equals: expected "setTimeout blocked" but got "setTimeout allowed"
